Add new Resources.js file and make exports in index.js
